move class registration onto own file
